% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importer.R
\name{MLHimport}
\alias{MLHimport}
\title{Import Microsoft Excel file with multiple line header as R tibble}
\usage{
MLHimport(path, lines, .args = list())
}
\arguments{
\item{path}{A string identifing the location of the Microsoft Excel file}

\item{lines}{The number of lines used as header in the Microsoft Excel file}

\item{.args}{Additional arguments to pass to readxl::read_excel function}
}
\value{
A tibble with a proper single line header
}
\description{
Import Microsoft Excel file with multiple line header as R tibble
}
\examples{
\dontrun{MLHimport("myExcelFileName.xlsx", lines = 3)}
}
